[2473] Fix expression dialog with implicit SI units referencing
Import the SI package via a NamespaceImport if an expression tries to use a standard unit without prefixing it with SI and SI is not already imported.

Bug: https://github.com/eclipse-syson/syson/issues/2473
