+++ doc/content/modules/user-manual/pages/release-notes/2026.9.0.adoc
@@ -126,6 +126,7 @@ The affected connections were missing their source and target and were therefore
 
 ** Fix expression creation and editing on `SuccessionAsUsage` elements.
 ** Fix the _Typed by_ property in the _Details_ view so that a `Usage` can only be typed by the corresponding kind of `Definition`.
+** Fix the expression dialog by automatically import the SI package via `NamespaceImport` when un-prefixed standard units are used in an expression.
 
 * In textual import/export:
